\name{msnbc323}
\docType{data}
\alias{msnbc323}
\title{Dataset: msnbc323}

\description{ A portion of the \code{msnbc} dataset containing 323
clickstream sequences. This version of the original dataset (David
Heckerman) was used in Melnykov (2014).\cr
There are 17 states representing the following categories:\cr
1: frontpage\cr
2: news\cr
3: tech\cr
4: local\cr
5: opinion\cr
6: on-air\cr
7: misc\cr
8: weather\cr
9: msn-news\cr
10: health\cr
11: living\cr
12: business\cr
13: msn-sports\cr
14: sports\cr
15: summary\cr
16: bbs\cr
17: travel\cr
}

\usage{data(msnbc323)}
\format{List of 323 numeric vectors representing categorical sequences.}
\source{Melnykov, V. (2014)}
\references{
Cadez, I., Heckerman, D., Meek, C., Smyth, P., White, S. (2003) Model-based clustering and visualization of navigation patterns on a web site, Data Mining and Knowledge Discovery, 399-424.

Melnykov, V. (2016) Model-Based Biclustering of Clickstream Data, Computational Statistics and Data Analysis, 93, 31-45.

Melnykov, V. (2016) ClickClust: An R Package for Model-Based Clustering of Categorical Sequences, Journal of Statistical Software, 74, 1-34.
}
\seealso{synth}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{dataset}% __ONLY ONE__ keyword per line
