\name{print.object}
\alias{print.EM}
\alias{summary.EM}
\alias{print.search}
\alias{summary.search}
\title{Functions for Printing or Summarizing Objects}
\description{
  \code{EM} and \code{search} classes  for printing and summarizing objects.
}
\usage{
\method{print}{EM}(x, ...)
\method{summary}{EM}(object, ...)
\method{print}{search}(x, ...)
\method{summary}{search}(object, ...)
}
\arguments{
  \item{x}{an object with the 'EM' (or 'search') class attributes.}
  \item{object}{an object with the 'EM' (or 'search') class attributes.}
  \item{\dots}{other possible options.}
}
\details{
  Some useful functions for printing and summarizing results.
}
\references{
Melnykov, V. (2014) Model-based biclustering of clickstream data, accepted by Computational Statistics & Data Analysis.}
\author{ Melnykov, V.}

\seealso{
  \code{click.EM}.
}

